The target co-ordinate parameters are now also properly written to the SQL as doubles, instead of converting them to integers again and therefore breaking the whole purpose.

+++ SAI-Editor/Forms/MainForm.cs
@@ -1669,10 +1669,10 @@ namespace SAI_Editor
             newSmartScript.target_param1 = XConverter.ToInt32(textBoxTargetParam1.Text);
             newSmartScript.target_param2 = XConverter.ToInt32(textBoxTargetParam2.Text);
             newSmartScript.target_param3 = XConverter.ToInt32(textBoxTargetParam3.Text);
-            newSmartScript.target_x = XConverter.ToInt32(textBoxTargetX.Text);
-            newSmartScript.target_y = XConverter.ToInt32(textBoxTargetY.Text);
-            newSmartScript.target_z = XConverter.ToInt32(textBoxTargetZ.Text);
-            newSmartScript.target_o = XConverter.ToInt32(textBoxTargetO.Text);
+            newSmartScript.target_x = XConverter.ToDouble(textBoxTargetX.Text);
+            newSmartScript.target_y = XConverter.ToDouble(textBoxTargetY.Text);
+            newSmartScript.target_z = XConverter.ToDouble(textBoxTargetZ.Text);
+            newSmartScript.target_o = XConverter.ToDouble(textBoxTargetO.Text);
 
             if (Settings.Default.GenerateComments && Settings.Default.UseWorldDatabase)
                 newSmartScript.comment = await CommentGenerator.Instance.GenerateCommentFor(newSmartScript, originalEntryOrGuidAndSourceType);
